feat: colorize service card chips
